added markup table and seeded values

+++ database/migrations/20190701082837_resources.js
@@ -2,7 +2,7 @@ exports.up = function(knex) {
   return knex.schema.createTable("resources", function(tbl) {
     tbl.increments("id").primary();
     tbl.string("name", 128)
-    tbl.integer("gold_value");
+    tbl.float("gold_value");
     tbl.string("type", 128);
   });
 };
